+++ common/views/components/Body/Body.js
@@ -2,7 +2,6 @@
 // TODO: Sync up types with the body slices and the components they return
 import dynamic from 'next/dynamic';
 import { classNames } from '../../../utils/classnames';
-import { london } from '../../../utils/format-date';
 import AsyncSearchResults from '../SearchResults/AsyncSearchResults';
 import { CaptionedImage } from '../Images/Images';
 import SpacingComponent from '../SpacingComponent/SpacingComponent';
@@ -24,55 +23,6 @@ import {
 } from '../../../services/prismic/html-serialisers';
 import { type Weight } from '../../../services/prismic/parsers';
 
-// TODO move this to service/
-// TODO flow
-function parseVenueTimesToOpeningHours(venue, daysInAdvance?: number) {
-  const data = venue.data;
-  const exceptionalOpeningHours = venue.data.modifiedDayOpeningTimes.map(
-    modified => {
-      const start =
-        modified.startDateTime &&
-        london(modified.startDateTime).format('HH:mm');
-      const end =
-        modified.startDateTime && london(modified.endDateTime).format('HH:mm');
-      const overrideDate =
-        modified.overrideDate && london(modified.overrideDate);
-      const overrideType = modified.type;
-      return {
-        overrideDate,
-        overrideType,
-        opens: start,
-        closes: end,
-      };
-    }
-  );
-  const weekArray = [
-    'monday',
-    'tuesday',
-    'wednesday',
-    'thursday',
-    'friday',
-    'saturday',
-    'sunday',
-  ];
-  return {
-    name: data.title,
-    openingHours: {
-      regular: weekArray.map(day => {
-        return {
-          dayOfWeek: day,
-          opens: data[day][0].startDateTime
-            ? london(data[day][0].startDateTime).format('HH:mm')
-            : null,
-          closes: data[day][0].endDateTime
-            ? london(data[day][0].endDateTime).format('HH:mm')
-            : null,
-        };
-      }),
-      exceptional: exceptionalOpeningHours,
-    },
-  };
-}
 const Map = dynamic(import('../Map/Map'), { ssr: false });
 
 type BodySlice = {|
@@ -219,9 +169,7 @@ const Body = ({ body, isDropCapped }: Props) => {
             {/* TODO parse in component */}
             {slice.type === 'venueTimes' && (
               <Layout8>
-                <OpeningHoursTable
-                  place={parseVenueTimesToOpeningHours(slice.value)}
-                />
+                <OpeningHoursTable place={slice.value} />
               </Layout8>
             )}
